bzdssm

+++ Socket-Common/src/main/java/core/Connector.java
@@ -50,7 +50,7 @@ public abstract class Connector implements Closeable, SocketChannelAdapater.OnCh
         }
 
         @Override
-        public ReceivePacket<?, ?> onArrivedNewPacket(byte type,int length) {
+        public ReceivePacket<?, ?> onArrivedNewPacket(byte type,long length) {
             switch (type){
                 case Packet.TYPE_MEMORY_BYTES:
                     return new BytesReceivePacket(length);
